Correct tests to account for auxiliary data changes
